src/main/java/device_management/Output_Devices/Computer.java
package device_management.Output_Devices;

import behaviours.IOutput;

import java.util.ArrayList;

public class Computer {
    private int ram;
    private int hddSize;
    private IOutput outputDevice;

//    public <I extends IOutput> Computer(int ram, int hddSize, I outputDevice) {
//        this.ram = ram;
//        this.hddSize = hddSize;
//        this.outputDevice = (outputDevice);
//    }

    public Computer(int ram, int hddSize, IOutput outputDevice) {
        this.ram = ram;
        this.hddSize = hddSize;
        this.outputDevice = outputDevice;
    }

    public int getRam() {
        return this.ram;
    }

    public int getHddSize() {
        return this.hddSize;
    }

    public IOutput getOutputDevice() {
        return this.outputDevice;
    }

    public String outputData(String data) {
        return this.outputDevice.outputData(data);
    }

    public void setOutputDevice(IOutput outputDevice) {
        this.outputDevice = outputDevice;
    }
}
